refactor: 型定義の整理とファイル構造の改善
- types/api.ts: API関連の型を集約
- types/domain.ts: ドメインモデルの型を集約（dailyReport.tsから改名）
- 重複していた型定義（Tag, DailyReportRequestType等）を削除
- GoogleUser型をutil.tsからdomain.tsに移動
- 全てのimportパスを更新
